% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{baseline_for_sim}
\alias{baseline_for_sim}
\title{Real baseline matrix corresponding to the simulated data in \code{simulation_data}.}
\format{
A \code{Matrix} with 10000 rows and 100 columns.
}
\usage{
baseline_for_sim
}
\description{
Real baseline matrix corresponding to the simulated data in \code{simulation_data}.
}
\keyword{datasets}
