% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init2.R
\name{Clean}
\alias{Clean}
\title{Delete all \code{.RData} files with given basename.
Can be used to clean the saved observation and baseline matrices.
Use with caution.}
\usage{
Clean(basename)
}
\arguments{
\item{basename}{A string.}
}
\value{
None
}
\description{
Delete all \code{.RData} files with given basename.
Can be used to clean the saved observation and baseline matrices.
Use with caution.
}
\examples{
Clean('33.0_observation_matrix_tmp')
Clean('observation_matrix_tmp')
}
