% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.r
\name{cmle}
\alias{cmle}
\title{Maximum likelihood estimate}
\usage{
cmle(data, n.cycles = 10, start = NULL, save.on.dir = TRUE)
}
\arguments{
\item{data}{.}

\item{n.cycles}{Number of time the conditional optimizers are called.}

\item{start}{Starting parameters. Functionality not yet implemented.}

\item{save.on.dir}{If TRUE, will save the inferred parameter in an .Rdata file}
}
\value{
A 2D 2x4 matrix with the estimated parameters.
}
\description{
Optimize the function \code{neg.log.like} recursively calling \code{clme1} and \code{cmle2}.
}
\examples{
cmle(data)
}
