% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.R
\name{cmle1}
\alias{cmle1}
\title{Compute and save on disk 2D Sparse matrices containing the observations recorded in \code{case.df} for each type.}
\usage{
cmle1(data, cpar, iterations = 10000)
}
\arguments{
\item{data}{.}

\item{cpar}{numeric.}

\item{iteration}{integer.}
}
\value{
2D 4x2 matrix.
}
\description{
Compute and save on disk 2D Sparse matrices containing the observations recorded in \code{case.df} for each type.
}
\examples{
cmle1(data, 1, 1000)
}
