% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.r, R/surveillance_utils.r
\name{compute.from.tab.baseline}
\alias{compute.from.tab.baseline}
\title{postcode.in.england}
\usage{
compute.from.tab.baseline(
  cylinder,
  observation.matrix,
  tab.baseline,
  postcode.locations
)

compute.from.tab.baseline(
  cylinder,
  observation.matrix,
  tab.baseline,
  postcode.locations
)
}
\arguments{
\item{x}{integer. number of cylinder samples.}

\item{postcode.field}{numeric.}
}
\description{
Check if postcode is in England.

Check if postcode is in England.
}
