% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.r
\name{CreateCylinders.delay}
\alias{CreateCylinders.delay}
\title{n.cylinders=10000 takes around 3 hours for the whole dataset, to end up with 300 non-empty cylinders
observation.matrix  and baseline.matrix have dimension 
This function scans the matrices
-1 in observation.matrix index means that we are excluding from week NA}
\usage{
CreateCylinders.delay(
  observation.matrix.typed,
  baseline.matrix.typed,
  observation.matrix.untyped,
  baseline.matrix.untyped,
  emmtype,
  week.range,
  n.cylinders = 1000,
  p.val.threshold = 0.05,
  coord.df = postcode2coord,
  GT = 24,
  size_factor = 1,
  rho
)
}
\arguments{
\item{observation.matrix.untyped}{A 2D matrix.}

\item{baseline.matrix.untyped}{A 2D matrix.}

\item{week.range}{A \code{numeric} vector to set the lower and upper limit to the heigh of the cylinders.}

\item{n.cylinders}{An \code{integer}, total number of drawn cylinders.}

\item{p.val.threshold}{A \code{numeric}. If the probability of observed exceedance is < \code{p.val.threshold}, flag the cylinder as anomalous.}

\item{size_factor}{A \code{numeric} multiplier to increase or reduce the cylinder heights and radia.}
}
\description{
n.cylinders=10000 takes around 3 hours for the whole dataset, to end up with 300 non-empty cylinders
observation.matrix  and baseline.matrix have dimension 
This function scans the matrices
-1 in observation.matrix index means that we are excluding from week NA
}
