% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.r
\name{CreateCylinders}
\alias{CreateCylinders}
\title{n.cylinders=10000 takes around 3 hours for the whole dataset, to end up with 300 non-empty cylinders
observation.matrix  and baseline.matrix have dimension 
This function scans the matrices
-1 in observation.matrix index means that we are excluding from week NA}
\usage{
CreateCylinders(
  observation.matrix,
  baseline.matrix,
  week.range,
  n.cylinders = 1000,
  rs = 0.1,
  p.val.threshold = 0.05,
  size_factor = 1
)
}
\arguments{
\item{n.cylinders}{(integer) number of proposed cylinders per week interval.}
}
\description{
n.cylinders=10000 takes around 3 hours for the whole dataset, to end up with 300 non-empty cylinders
observation.matrix  and baseline.matrix have dimension 
This function scans the matrices
-1 in observation.matrix index means that we are excluding from week NA
}
