% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init2.R
\name{EmmtypeFactor.delay_}
\alias{EmmtypeFactor.delay_}
\title{At a given week, a fraction lambda_untyped \approx 0.6 of all cases are not typed.
the baselines e.g. are as follows:
- for the emmtype 33.0: (1-lambda_untyped) * lambda_33.0 * lambda_t * lambda_geo
- for the entyped: lambda_untyped * lambda_t * lambda_geo}
\usage{
EmmtypeFactor.delay_(case.file, starting.week, n.weeks)
}
\description{
At a given week, a fraction lambda_untyped \approx 0.6 of all cases are not typed.
the baselines e.g. are as follows:
- for the emmtype 33.0: (1-lambda_untyped) * lambda_33.0 * lambda_t * lambda_geo
- for the entyped: lambda_untyped * lambda_t * lambda_geo
}
