% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_utils.r
\name{f_radia_and_heights}
\alias{f_radia_and_heights}
\title{Find optimal radia}
\usage{
f_radia_and_heights(baseline.matrix, heigths = 1:100)
}
\arguments{
\item{baseline.matrix}{A \code{Matrix} enconding the baseline.}

\item{heigths}{\code{integer}.}
}
\value{
A \code{Matrix}.
}
\description{
This function compute the optimal cylinder radia, such that the corresponding
cylinders contain one event in average 
for the chosen N heights (\code{heights}) and a given \code{baseline.matrix}.
It returns an Nx2 \code{Matrix} whose first column contains the heights and 
and the second column contains the corresponding radia.
}
\examples{
radia_and_heights = f_radia_and_heights(baseline.matrix, 1:10)
}
