% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_radii_and_heights.r
\name{f_radii_and_heights2}
\alias{f_radii_and_heights2}
\title{Find cylinders' radii and heights}
\usage{
f_radii_and_heights2(heigths, baseline.matrix, target.n.event, rho)
}
\arguments{
\item{heigths}{\code{integer}.}

\item{baseline.matrix}{A \code{Matrix} encoding the baseline.}

\item{target.n.event}{An \code{integer}, representing the desired expected number of events in each cylinder. It requires a baselineline matrix to be passed.}

\item{rho}{A \code{numeric}. It is the maximum}
}
\value{
A \code{Matrix}.
}
\description{
This function returns cylinder radii and heights, such that the corresponding
cylinder volumes have determined characteristics.
For the chosen N heights (\code{heights}),
it returns an Nx2 \code{Matrix} whose first column contains the heights and 
and the second column contains the corresponding radii.
It works in two modes: 1) if \code{baseline.matrix} and \code{target.n.event} are provided (but \code{rho} is not),
the function computes the volume (such that each cylinder contains around \code{target.n.event}
events given the baseline); this is the mode used in Cavallaro, M. et al (2022) PLoS Comput Biol 18(11): e1010726
https://doi.org/10.1371/journal.pcbi.1010726 .
2) if \code{rho} is provided (but  \code{baseline.matrix} and \code{target.n.event} are not)
the functions computes the largest volume as $pi * rho^2$. All other radii
are computed with heights from \code{heights} such the the cylinder volumes are constant.
}
\examples{
radii_and_heights = f_radii_and_heights2(1:16, baseline.matrix, 100)
radii_and_heights = f_radii_and_heights2(1:16, rho=1.6446127)
}
