% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_radii_and_heights.r
\name{f_radii_and_heights_}
\alias{f_radii_and_heights_}
\title{This funcion is analolgous to f_radii_and_heights2, but is takes a tabular baseline
instead of a baseline matrix.}
\usage{
f_radii_and_heights_(baseline.tab, heights = 1:100)
}
\description{
This funcion is analolgous to f_radii_and_heights2, but is takes a tabular baseline
instead of a baseline matrix.
}
