% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rancovr.r
\docType{package}
\name{foo}
\alias{foo}
\title{rancovr: Cluster detection in R with RAndom Neighbourhood COVeRing}
\description{
rancovr detects statio-temporal clusters against a baseline of
sparse Poisson point events using
the random neighbourhood covering algorithm.
}
\section{Foo functions}{

The foo functions ...
}

