% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{GB_region_boundaries}
\alias{GB_region_boundaries}
\title{GB boundaries.}
\format{
Simple feature collection.
}
\source{
\url{https://osdatahub.os.uk/}
}
\usage{
GB_region_boundaries
}
\description{
A dataset containing the boundaries of the Euro constituencies of all Great Britain,
Coordinate reference systems (CRS) EPSG code 4326, which uses units of longitude and latitude
on the  World Geodetic System 1984 (WGS84) ellipsoid.
This information is license under the Open Government Licence v3.0, see \url{https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}.
}
\keyword{datasets}
