% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.r
\name{lambda}
\alias{lambda}
\title{Lambda
 
  \lambda(t) = a + b + b \sin(c\,2 \pi/365 + x 2 \,\pi / 365)}
\usage{
lambda(x, params)
}
\arguments{
\item{x}{vector of time points.}

\item{params}{vector of parameters of length = 4.}
}
\value{
a vector.
}
\description{
Lambda
 
  \lambda(t) = a + b + b \sin(c\,2 \pi/365 + x 2 \,\pi / 365)
}
\examples{
lambda(c(1,2,3,4,5,6), c(1,1,1,1) )
}
