% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{make_circle}
\alias{make_circle}
\title{centers: the data frame of centers with ID
radius: radius measured in kilometer}
\usage{
make_circle(x0, y0, radius, n.points = 100)
}
\description{
centers: the data frame of centers with ID
radius: radius measured in kilometer
}
