% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.r
\name{neg.log.like}
\alias{neg.log.like}
\title{Negative log-likelihood}
\usage{
neg.log.like(params, data)
}
\arguments{
\item{params}{vector of parameters of length = 4.}

\item{x}{vector of time data.}
}
\value{
a numeric
}
\description{
$$
  l(\lambda) = \log L(\lambda) = - \int_0^T\lambda(x) dx + \sum_{i=1}^n\log \lambda(x_i)
$$
}
\examples{
lambda(c(1,2,3,4,5,6), c(1,1,1,1) )
}
