% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{plotCircle}
\alias{plotCircle}
\title{centers: the data frame of centers with ID
radius: radius measured in kilometer}
\usage{
plotCircle(cylinders, n.points = 100, add = T, ...)
}
\description{
centers: the data frame of centers with ID
radius: radius measured in kilometer
}
