% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcode.data}
\alias{postcode.data}
\title{A dataset containing the population and the centroid coordinates of all UK postcodes.}
\format{
A data frame with 1048575 rows and 7 variables:
\describe{
  \item{postcode}{}
  \item{Total}{}
  \item{latitude}{}
  \item{longitude}{}
}
}
\source{
\url{https://www.ons.gov.uk}
}
\usage{
postcode.data
}
\description{
A dataset containing the population and the centroid coordinates of all UK postcodes.
}
\keyword{datasets}
