% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{postcode.in.england}
\alias{postcode.in.england}
\title{postcode.in.england}
\usage{
postcode.in.england(x, postcode.field = "postcode")
}
\arguments{
\item{x}{integer. number of cylinder samples.}

\item{postcode.field}{numeric.}
}
\value{
A \code(data.frame).
}
\description{
Check if postcode is in England.
}
