% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postcode.to.location.and.population}
\alias{postcode.to.location.and.population}
\title{postcode.to.location.and.population}
\usage{
postcode.to.location.and.population(x, postcodes, postcode.field = "postcode")
}
\arguments{
\item{x}{integer. number of cylinder samples.}

\item{postcodes}{numeric.}

\item{postcode.field}{numeric.}
}
\value{
A \code(data.frame).
}
\description{
Find geo coordinates and population from postcode data frame.
}
