% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postcode.to.location2}
\alias{postcode.to.location2}
\title{postcode.to.location2}
\usage{
postcode.to.location2(x, postcode.field = "postcode")
}
\arguments{
\item{x}{integer. number of cylinder samples.}

\item{postcode.field}{numeric.}
}
\value{
A \code(data.frame).
}
\description{
postcode.to.location2
}
