% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{postcode.to.location3}
\alias{postcode.to.location3}
\title{postcode.to.location3}
\usage{
postcode.to.location3(x)
}
\arguments{
\item{x}{integer. number of cylinder samples.}
}
\value{
A \code(data.frame).
}
\description{
postcode.to.location3
}
