% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postcode.to.region}
\alias{postcode.to.region}
\title{postcode.to.region}
\usage{
postcode.to.region(x, Area2Region_list)
}
\arguments{
\item{x}{integer. number of cylinder samples.}

\item{Area2Region_list}{numeric.}
}
\value{
A \code(data.frame)
}
\description{
postcode.to.region
}
