% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_utils.r
\name{rcylinder}
\alias{rcylinder}
\title{rcylinders}
\usage{
rcylinder(n, X.range, Y.range, time.range, border = NULL, rs = 0.5, a = 0)
}
\arguments{
\item{n}{integer. number of cylinder samples.}

\item{X.range}{numeric.}

\item{Y.range}{numeric.}

\item{time.range}{numeric}

\item{border}{NULL or numeric. It is the distance from the boundary.}

\item{rs}{numeric. It is a reference size.}
}
\value{
A \code(data.frame).
}
\description{
Compute random cylinder coordinates.
}
\examples{
cylinders=rcylinder(10, c(-2,3), c(-10,10)), c(1, 100) )
}
