% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.r
\name{Simulate}
\alias{Simulate}
\title{Simulate an observation matrix}
\usage{
Simulate(
  population,
  time.factor,
  total.average,
  save.baseline.matrix = F,
  n.unknown.time = 0
)
}
\arguments{
\item{population}{A \code{numeric} vector.}

\item{time.factor}{\code{numeric} vector.}

\item{total.average}{A \code{numeric}.}

\item{save.baseline.matrix}{A \code{logical}.}

\item{n.unknown.time}{A \code{numeric}. Represents the baseline for cases that have unknown date report.}
}
\value{
A \code{sparseMatrix} observation matrix.
}
\description{
Given  vectors of length M and N, representing spatial 
and temporal factors (e.g., population density per location and sesonal trends, respectively),
this function computes a MxN baseline matrix and simulate a Poisson point process.
}
\examples{
sim = Simulate(spatial.factor, time.factor, total.average)
}
