% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{simulation_data}
\alias{simulation_data}
\title{Simulation data.}
\format{
A data frame with 5028 rows and 9 variables:
\describe{
  \item{week}
  \item{postcode}{}
  \item{population}{}
  \item{sim}{}
  \item{type}{}
  \item{latitude}{}
  \item{longitude}{}
  \item{y}{} 
  \item{x}{}  
}
}
\source{
\url{https://www.ons.gov.uk}
}
\usage{
simulation_data
}
\description{
A dataset containing.
}
\keyword{datasets}
