% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.r
\name{tabulated.baseline}
\alias{tabulated.baseline}
\title{tabulated.baseline}
\usage{
tabulated.baseline(case.df, date.time.field = "week")
}
\arguments{
\item{case.df}{A \code{data.frame} of events.}

\item{date.time.field}{A \code{character} string.}
}
\description{
tabulate the baseline intensity function.
}
