% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.r
\name{tabulated.baseline2}
\alias{tabulated.baseline2}
\title{tabulated.baseline}
\usage{
tabulated.baseline2(case.df, date.time.field = "week")
}
\arguments{
\item{x}{integer. number of cylinder samples.}

\item{postcode.field}{numeric.}
}
\description{
tabulate the baseline intensity function.
}
