% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init2.R
\name{TimeFactor}
\alias{TimeFactor}
\title{Returns a vector represententing the temporal component of the baseline.}
\usage{
TimeFactor(
  case.df,
  save.on.dir = TRUE,
  get.from.dir = FALSE,
  date.time.field = "week",
  parameters = NULL,
  n.iterations = 20
)
}
\arguments{
\item{case.df}{A data.frame}

\item{save.on.dir}{A logical}

\item{get.from.dir}{logical}

\item{date.time.field}{A string}

\item{n.iterations}{An integer}
}
\value{
A vector
}
\description{
Returns a vector represententing the temporal component of the baseline.
}
\examples{
time.factor = TimeFactor(case.df)
}
