% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.r
\name{UpdateTimeFactor}
\alias{UpdateTimeFactor}
\title{Update Time factor}
\usage{
UpdateTimeFactor(
  case.df,
  save.on.dir = TRUE,
  get.from.dir = FALSE,
  date.time.field = "week",
  parameters = NULL,
  n.iterations = 20
)
}
\arguments{
\item{case.df}{A \code{data.frame} containing the events.}

\item{save.on.dir}{\code{logical}. If TRUE then the vector is saved in `timefactor.RData` file.}

\item{get.from.dir}{\code{logical}. If TRUE then the vector is obtained from the `timefactor.RData` file.}

\item{date.time.field}{A \code{character} string.}

\item{n.iterations}{An \code{integer}.}
}
\value{
A vector
}
\description{
Returns a vector representing the temporal component of the baseline.
}
\examples{
time.factor = TimeFactor(case.df)
}
