% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.r
\name{UpgradeBaselineMatrix}
\alias{UpgradeBaselineMatrix}
\title{Upgrade baseline matrix}
\usage{
UpgradeBaselineMatrix(
  previous.baseline.matrix,
  case.df.old,
  case.df.new,
  save.on.dir = FALSE,
  date.time.field = "week",
  postcode.field = "postcode",
  n.iterations = 10
)
}
\arguments{
\item{save.on.dir}{\code{logical}. If TRUE then the vector is saved in `baseline_matrix.RData` file.}

\item{date.time.field}{A \code{character} string.}

\item{postcode.field}{A \code{character} string.}
}
\value{
A 2D matrix
}
\description{
Create a new baseline matrix to match the new observation matrix and makes new fit
for the temporal factor.
}
