% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_utils.r
\name{warning.score}
\alias{warning.score}
\title{Compute the warning score of a case (location and date).}
\usage{
warning.score(case, cylinders, date.time.field = "week")
}
\arguments{
\item{case}{A \code{numeric}; .}

\item{cylinders}{A \code{data.frame} object encoding the events.}

\item{date.time.field}{A \code{character}.}
}
\value{
A \code{numeric}.
}
\description{
Compute the warning score of a case -- defined by its location (coordinates) 
and data -- for a given set of cylinders (computed with \code{}).
It can be used in \code{apply}.
}
\examples{
warning.scores = apply(cases, 1, warning.score, cylinders)
}
