% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_utils.r
\name{warning.score2}
\alias{warning.score2}
\title{Compute the warning score of a case (location and date).}
\usage{
warning.score2(case, TT, cylinders)
}
\arguments{
\item{case}{A \code{numeric}; .}

\item{TT}{An \code{integer}.}

\item{cylinders}{A \code{data.frame} object encoding the events.}
}
\value{
A \code{numeric} vector of size two.
}
\description{
Compute the warning score of a case -- defined by its location (coordinates) 
and data -- for a given set of cylinders (computed with \code{CreateCylinders}).
It can be used in \code{apply}.
}
\examples{
tmp = apply(cases, 1, warning.score, TT, cylinders)
warning.score = tmp[1,]
}
