% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_utils.r
\name{cmle2}
\alias{cmle2}
\title{Optimize the function \code{neg.log.like} conditioned on having the first three parameters fixed.}
\usage{
cmle2(data, cpar)
}
\arguments{
\item{data}{.}

\item{cpar}{numeric.}
}
\value{
numeric.
}
\description{
Optimize the function \code{neg.log.like} conditioned on having the first three parameters fixed.
}
\examples{
cmle2(data, c(1,1,1))
}
