% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_utils.r
\name{compute.from.tab.baseline}
\alias{compute.from.tab.baseline}
\title{Compute exceedance probability in a cylinder.}
\usage{
compute.from.tab.baseline(
  cylinder,
  observation.matrix,
  tab.baseline,
  postcode.locations
)
}
\arguments{
\item{cylinder}{}

\item{observation.matrix}{A \code{sparseMatrix} object encoding the events.}

\item{tab.baseline}{An \code{expand.grid} tab encoding the baseline.}

\item{postcode.locations}{A \code{data.frame} that maps the rows of \code{observation.matrix} to geographical coordinates.}
}
\value{
A \code{numeric} vector of dimension 3.
}
\description{
A cylinder is defined by the circle coordinated (say, x,y, and radius) and lower and upper height limits (aay, t.low and t.upp, respectively).
For a given cylinder, this function computes the number of observed events (\code{n_cases}) in the cylinder according to
\code{observation.matrix}, the expected number \code{mu} of events according the Poisson point model (with intensity
defined in \code{tab.baseline}), and the probability that .
The function returns \code{c(n_cases, mu, p.val)}.
}
\examples{
exceedance=compute(c(x,y,rho,t.low,t.upp), observation.matrix, tab.baseline, postcode.locations)
}
