% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.r
\name{CreateBaselineMatrix}
\alias{CreateBaselineMatrix}
\title{Create and save a baseline matrix}
\usage{
CreateBaselineMatrix(
  case.df,
  save.on.dir = FALSE,
  date.time.field = "week",
  postcode.field = "postcode"
)
}
\arguments{
\item{case.df}{A \code{data.frame} of events.}

\item{save.on.dir}{\code{logical}. If TRUE then the vector is saved in `baseline_matrix.Rdata` file.}

\item{date.time.field}{A \code{character} string.}

\item{postcode.field}{A \code{character} string.}
}
\value{
A \code{Matrix}.
}
\description{
Compute and save on disk a 2D dense matrix representing the spatial component of the
baseline.
}
\examples{
baseline.matrix = CreateBaselineMatrix(case.df)
baseline.matrix = CreateBaselineMatrix(case.df, date.time.field = 'SAMPLE_DT_numeric', postcode.field = 'Patient Postcode')
}
