% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init2.R
\name{CreateBaselineMatrix}
\alias{CreateBaselineMatrix}
\title{Create and save a baseline matrix}
\usage{
CreateBaselineMatrix(
  case.df,
  save.on.dir = FALSE,
  date.time.field = "week",
  postcode.field = "postcode"
)
}
\arguments{
\item{case.df}{A \code{data.frame}.}

\item{date.time.field}{A character string.}

\item{postcode.field}{A character string.}
}
\value{
A matrix
}
\description{
Compute and save on disk a 2D dense matrix containing the spatial component of the
baseline.
}
\examples{
baseline.matrix = CreateBaselineMatrix(case.df)
baseline.matrix = CreateBaselineMatrix(case.df, date.time.field = 'SAMPLE_DT_numeric', postcode.field = 'Patient Postcode')
}
