% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.r
\name{CreateCylinders.delay}
\alias{CreateCylinders.delay}
\title{n.cylinders=10000 takes around 3 hours for the whole dataset, to end up with 300 non-empty cylinders
observation.matrix  and baseline.matrix have dimension 
This function scans the matrices
-1 in observation.matrix index means that we are excluding from week NA}
\usage{
CreateCylinders.delay(
  observation.matrix.typed,
  baseline.matrix.typed,
  observation.matrix.untyped,
  baseline.matrix.untyped,
  emmtype,
  week.range,
  n.cylinders = 1000,
  rs = 0.1,
  p.val.threshold = 0.05,
  coord.df = postcode2coord,
  size_factor = 1
)
}
\arguments{
\item{observation.matrix.untyped}{A 2D matrix.}

\item{baseline.matrix.untyped}{A 2dD matrix.}

\item{week.range}{week range.}

\item{n.cylinders}{(integer) number of proposed cylinders per week interval.}

\item{rs}{A numeric.}

\item{p.val.threshold}{A numeric.}

\item{size_factor}{An integer.}
}
\description{
n.cylinders=10000 takes around 3 hours for the whole dataset, to end up with 300 non-empty cylinders
observation.matrix  and baseline.matrix have dimension 
This function scans the matrices
-1 in observation.matrix index means that we are excluding from week NA
}
