% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init2.R
\name{CreateObservationMatrices}
\alias{CreateObservationMatrices}
\title{Compute and save on disk 2D Sparse matrices containing the observations recorded in \code{case.df} for each type.}
\usage{
CreateObservationMatrices(
  case.df,
  types = NA,
  date.time.field = "week",
  postcode.field = "postcode"
)
}
\arguments{
\item{case.df}{A data frame.}

\item{types}{A vector of strings.}

\item{date.time.field}{A string.}

\item{postcode.field}{A string.}
}
\value{
None
}
\description{
Compute and save on disk 2D Sparse matrices containing the observations recorded in \code{case.df} for each type.
}
\examples{
CreateObservationMatrices(case.df)
CreateObservationMatrices(case.df, types=c('1.0', '33.0'), date.time.field = 'SAMPLE_DT_numeric', postcode.field = 'Patient Postcode')
}
