% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.r
\name{PostcodeMap}
\alias{PostcodeMap}
\title{Map postcodes to coordinates
Returns and save is 'postcode2coord.Rdata' a data frame that maps the postcodes included in \code{rownames(matrix)} to geographical coordinates.
\code{matrix} can be a \code{Matrix} or a \code{sparseMatrix} object storing baseline or observation data.
Requires the all postcode data tabulated in a \code{data.frame} called \code{postcode.data} available in the workspace.}
\usage{
PostcodeMap(matrix, postcode.field = "postcode")
}
\arguments{
\item{matrix}{A \code{Matrix} or \code{sparseMatrix}.}
}
\value{
A \code{data.frame}.
}
\description{
Map postcodes to coordinates
Returns and save is 'postcode2coord.Rdata' a data frame that maps the postcodes included in \code{rownames(matrix)} to geographical coordinates.
\code{matrix} can be a \code{Matrix} or a \code{sparseMatrix} object storing baseline or observation data.
Requires the all postcode data tabulated in a \code{data.frame} called \code{postcode.data} available in the workspace.
}
\examples{
postcode2coord = PostcodeMap(observation.matrix)
}
