% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance_utils.r
\name{rcylinder}
\alias{rcylinder}
\title{Draw random cylinder coordinates}
\usage{
rcylinder(
  n.cylinders,
  observation.matrix,
  time.range,
  radia_and_heights,
  postcode2coord
)
}
\arguments{
\item{n.cylinders}{An \code{integer}; the number of cylinders to draw.}

\item{observation.matrix}{A \code{sparseMatrix} object enconding the events.}

\item{time.range}{An \code{integer} vector.}

\item{radia_and_heights}{A \code{Matrix}.}

\item{postcode2coord}{A \code{data.frame} that maps the rows of \code{observation.matrix} to geographical coordinates.}
}
\value{
A \code(data.frame).
}
\description{
Find the coordinates (centers and height limits) of cylinders
that contain events defined in \code{observation.matrix}, with radia and heights
given in \code{radia_and_heights}.
}
\examples{
cylinders=rcylinder(10, observation.matrix, time.range, radia_and_heights, postcode2coord)
}
