% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.r
\name{UpdateBaselineMatrix}
\alias{UpdateBaselineMatrix}
\title{Update baseline matrix}
\usage{
UpdateBaselineMatrix(
  previous.baseline.matrix,
  recent.case.df,
  save.on.dir = FALSE,
  date.time.field = "week",
  postcode.field = "postcode"
)
}
\arguments{
\item{save.on.dir}{\code{logical}. If TRUE then the vector is saved in `baseline_matrix.Rdata` file.}

\item{date.time.field}{A \code{character} string.}

\item{postcode.field}{A \code{character} string.}
}
\value{
A 2D matrix
}
\description{
Create a new baseline matrix to match the new observation matrix.
This function is very similar to \code{UpgradeBaselineMatrix}
does not perform fit if time.factor parameters were saved,
only create a baseline matrix to match the new observation matrix.
}
