% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.r
\name{UpdateObservationMatrices}
\alias{UpdateObservationMatrices}
\title{Upgrade observation matrix}
\usage{
UpdateObservationMatrices(
  case.df.old,
  case.df.new,
  types = NULL,
  date.time.field = "week",
  postcode.field = "postcode"
)
}
\arguments{
\item{types}{\code{NULL} or a \code{character} vector.}

\item{date.time.field}{A \code{character} string.}

\item{postcode.field}{A \code{character} string.}
}
\description{
Create a new observation matrix and overwrite the old one
}
