% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.r
\name{UpgradeBaselineMatrix}
\alias{UpgradeBaselineMatrix}
\title{Upgrade baseline matrix}
\usage{
UpgradeBaselineMatrix(
  case.df.new,
  case.df.old,
  save.on.dir = FALSE,
  date.time.field = "week",
  postcode.field = "postcode"
)
}
\arguments{
\item{date.time.field}{A character string.}

\item{postcode.field}{A character string.}
}
\value{
A 2D matrix
}
\description{
Create a new baseline matrix to match the new observation matrix and makes new fit
}
